Move face_model class to core
