+++ src/face_projection/core.py
@@ -1,14 +1,34 @@
 __all__ = ["Warper"]
 
+from pathlib import Path
 from typing import Optional
 
 import cv2
+import h5py
 import mediapipe as mp
 import numpy as np
 from meshpy import triangle
 
 from . import consts
-from .face_model import FaceModel
+
+
+class FaceModel:
+    """Face model class.
+
+    The underlying UV model is based on the canonical face model by google:
+        Link: https://github.com/google/mediapipe/blob/master/mediapipe/modules/face_geometry/data/canonical_face_model.obj
+        Image: https://github.com/google/mediapipe/blob/master/mediapipe/modules/face_geometry/data/canonical_face_model_uv_visualization.png
+
+    """
+
+    def __init__(self) -> None:
+        data_file = h5py.File(Path(__file__).parent / "face_model.h5", "r")
+
+        self.points = np.array(data_file["points"])
+        self.triangles = np.array(data_file["triangles"])
+        self.facets = np.array(data_file["facets"])
+        self.masking = np.array(data_file["masking_canonical"])
+        data_file.close()
 
 
 class Warper:
